backend/requirements.txt
numpy==1.26.4
pandas==2.1.4
joblib==1.3.2
scikit-learn==1.4.2
pydantic==2.6.4
tensorflow==2.15.0
uvicorn
fastapi
python-dotenv